More work converting Stellaris script to YAML
leader classes could be just words with spaces,
so go back and form a list from them
so we can parse them
